// src/types/infoTypes.ts
export enum InfoType {
    Id = 'id',
    Name = 'name',
    Description = 'description',
    Link = 'link',
    Type = 'type',
    Updates = 'updates',
    Project = 'project',
    Projects = 'projects',
    Environment = 'environment',
    Environments = 'environments',
    Repository = 'repository',
    Commits = 'commits',
    Version = 'version',
    Versions = 'versions',
    Output = 'output',
    Timestamp = 'timestamp',
    Git = 'git',
}